Move transformSVGAtt to utils Because it will be easier to test and it does not touch `this`
